PROD-1239 adds query param to GET experience to include meta (#4335)
