+++ clients/fides-js/src/services/fides/api.ts
@@ -41,6 +41,7 @@ export const fetchExperience = async (
     has_config: "true",
     systems_applicable: "true",
     include_gvl: "true",
+    include_meta: "true",
   };
   if (fidesUserDeviceId) {
     params.fides_user_device_id = fidesUserDeviceId;
