+++ clients/privacy-center/cypress/e2e/consent-banner-tcf.cy.ts
@@ -80,13 +80,18 @@ describe("Fides-js TCF", () => {
           experience: experience.items[0],
         });
       });
+      cy.intercept("PATCH", `${API_URL}${FidesEndpointPaths.NOTICES_SERVED}`, {
+        fixture: "consent/notices_served_tcf.json",
+      }).as("patchNoticesServed");
     });
-    cy.intercept("PATCH", `${API_URL}${FidesEndpointPaths.NOTICES_SERVED}`, {
-      fixture: "consent/notices_served_tcf.json",
-    }).as("patchNoticesServed");
   });
 
   describe("banner appears when it should", () => {
+    beforeEach(() => {
+      cy.intercept("PATCH", `${API_URL}${FidesEndpointPaths.NOTICES_SERVED}`, {
+        fixture: "consent/notices_served_tcf.json",
+      }).as("patchNoticesServed");
+    });
     const setAllTcfToValue = (
       experience: PrivacyExperience,
       value: UserConsentPreference | undefined
@@ -197,6 +202,9 @@ describe("Fides-js TCF", () => {
           experience: experience.items[0],
         });
       });
+      cy.intercept("PATCH", `${API_URL}${FidesEndpointPaths.NOTICES_SERVED}`, {
+        fixture: "consent/notices_served_tcf.json",
+      }).as("patchNoticesServed");
     });
     it("can render purposes in the initial layer as a stack", () => {
       cy.get("div#fides-banner").within(() => {
@@ -243,6 +251,9 @@ describe("Fides-js TCF", () => {
           experience: experience.items[0],
         });
       });
+      cy.intercept("PATCH", `${API_URL}${FidesEndpointPaths.NOTICES_SERVED}`, {
+        fixture: "consent/notices_served_tcf.json",
+      }).as("patchNoticesServed");
       cy.get("#fides-modal-link").click();
     });
 
@@ -875,9 +886,8 @@ describe("Fides-js TCF", () => {
                   data_sales: true,
                   tracking: false,
                 });
-                expect(args[1]).to.equal(
-                  "CP0JloAP0JloAGXABBENATEAAAAAAAAAAAAAAAAAAAAA.IABE,1~"
-                );
+                // the TC str is dynamically updated upon save preferences with diff timestamp, so we do a fuzzy match
+                expect(args[1]).to.contain(".IABE,1~");
                 expect(args[2]).to.deep.equal(privacyExperience.items[0]);
               });
               // timeout means API call not made, which is expected
