+++ clients/privacy-center/types/api/index.ts
@@ -16,7 +16,6 @@ export type { EmbeddedLineItem } from "./models/EmbeddedLineItem";
 export type { EmbeddedVendor } from "./models/EmbeddedVendor";
 export { EnforcementLevel } from "./models/EnforcementLevel";
 export type { ExperienceConfigResponse } from "./models/ExperienceConfigResponse";
-export type { ExperienceMeta } from "./models/ExperienceMeta";
 export type { Identity } from "./models/Identity";
 export type { IdentityVerificationConfigResponse } from "./models/IdentityVerificationConfigResponse";
 export type { LastServedConsentSchema } from "./models/LastServedConsentSchema";
@@ -43,5 +42,4 @@ export type { TCFVendorConsentRecord } from "./models/TCFVendorConsentRecord";
 export type { TCFVendorLegitimateInterestsRecord } from "./models/TCFVendorLegitimateInterestsRecord";
 export type { TCFVendorRelationships } from "./models/TCFVendorRelationships";
 export type { TCFVendorSave } from "./models/TCFVendorSave";
-export { TCMobileData } from "./models/TCMobileData";
 export { UserConsentPreference } from "./models/UserConsentPreference";
